fix(dev): host the EntryCard preview gallery in a 'use dom' document
Post-pivot EntryCard renders entry HTML through a lowercase <div>, so the dev route crashed in the native RN tree (no div host). Move the samples into a 'use dom' document mirroring the reader; the ThemePicker stays native and threads its id in.
